requirements.txt
bcrypt==3.1.7
cffi==1.14.4
cryptography==2.8
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.3.0
python-dotenv==0.15.0
six==1.15.0
